pyproject.toml
[tool.poetry]
name = "digirent"
version = "0.1.0"
description = ""
authors = ["Shaibu Shaibu <s.shaibu.jnr@gmail.com>"]

[tool.poetry.scripts]
create_admin_user = "digirent.script:create_admin_user"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.55.1"
httpx = "^0.13.3"
uvicorn = "^0.11.5"
sqlalchemy = "^1.3.17"
psycopg2 = "^2.8.5"
alembic = "^1.4.2"
sqlalchemy_utils = "^0.36.6"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
python-multipart = "^0.0.5"
pyjwt = "^1.7.1"
dependency_injector = "^3.15.6"
pytz = "^2020.1"
aiofiles = "^0.5.0"
stripe = "^2.48.0"
sendgrid = "^6.3.1"
itsdangerous = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest_cov = "^2.10.1"
pytest-mock = "^3.1.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
